Lower PWM freq. on cal. and add setting for T.res.
Time resolution can now be set from h-file. The setting will be adjusted
in the setup function. Due to the auto-calibration feature, this is less
important, but still a neat feature.

The PWM frequency change concerns the sweeping of the PWM outputs to
signal ongoing calibration. Previously, the speed was a bit high.